requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
django-rest-knox==4.2.0
djangorestframework==3.14.0
importlib-metadata==5.0.0
json-spec==0.11.0
jsoncomment==0.4.2
pycparser==2.21
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
zipp==3.10.0
